+++ custom_components/metlink_explorer/sensor.py
@@ -241,7 +241,7 @@ class MetlinkRouteSensor(CoordinatorEntity, SensorEntity):
             "name": f"{transportation_name} Route {route_short_name}",
             "manufacturer": "Metlink",
             "model": transportation_name,
-            "sw_version": "0.5.0",
+            "sw_version": "0.5.1",
         }
 
     @property
@@ -335,7 +335,7 @@ class MetlinkDirectionSensor(CoordinatorEntity, SensorEntity):
             "name": f"{transportation_name} Route {route_short_name}",
             "manufacturer": "Metlink",
             "model": transportation_name,
-            "sw_version": "0.5.0",
+            "sw_version": "0.5.1",
         }
 
     @property